requirements.txt
alembic==1.8.0
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcolz-zipline==1.2.4
Bottleneck==1.3.5
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.1
coveralls==3.3.1
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
distlib==0.3.4
docopt==0.6.2
empyrical-reloaded==0.5.8
exchange-calendars==3.6.3
executing==0.8.3
filelock==3.7.1
fonttools==4.34.3
greenlet==1.1.2
h5py==3.7.0
ibapi==9.81.1.post1
idna==3.3
inflection==0.5.1
intervaltree==3.1.0
ipython==8.4.0
iso3166==2.0.2
iso4217==1.9.20220401
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.3
korean-lunar-calendar==0.2.1
Logbook==1.5.3
lru-dict==1.1.7
lxml==4.9.1
mailjet-rest==1.3.4
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
memoization==0.4.0
more-itertools==8.13.0
multipledispatch==0.6.0
multitasking==0.0.11
Nasdaq-Data-Link==1.0.2
networkx==2.8.4
numexpr==2.8.3
numpy==1.23.0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.10.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
polling==0.3.2
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyfolio-reloaded==0.9.4
Pygments==2.12.0
pyluach==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-interface==1.6.1
pytz==2022.1
requests==2.28.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
setuptools-scm==7.0.4
singleton-decorator==1.0.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
stack-data==0.3.0
statsmodels==0.13.2
TA-Lib==0.4.24
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tox==3.25.1
tox-gh-actions==2.9.1
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.10
virtualenv==20.15.1
wcwidth==0.2.5
yfinance==0.1.72
pandas_datareader==0.10.0
